Move entity factory into data service
