+++ libs/data/src/index.ts
@@ -1,3 +1,5 @@
 export * from './lib/data.service';
 export * from './lib/data.module';
-export * from './lib/truncate-tables';
+export * from './util/truncate-tables';
+export * from './util/entity.factory';
+export * from './util/mock-class';
