+++ libs/data/src/util/mock-class.ts
