--- libs/factory/README.md
@@ -1,7 +1,0 @@
-# factory
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test factory` to execute the unit tests via [Jest](https://jestjs.io).
