--- libs/factory/project.json
@@ -1,23 +1,0 @@
-{
-  "root": "libs/factory",
-  "sourceRoot": "libs/factory/src",
-  "projectType": "library",
-  "targets": {
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["libs/factory/**/*.ts"]
-      }
-    },
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/libs/factory"],
-      "options": {
-        "jestConfig": "libs/factory/jest.config.js",
-        "passWithNoTests": true
-      }
-    }
-  },
-  "tags": []
-}
