--- libs/factory/src/index.ts
@@ -1,2 +1,0 @@
-export * from './lib/entity.factory';
-export * from './lib/mock-class';
