--- libs/factory/src/lib/client.ts
@@ -1,4 +1,0 @@
-import { PrismaClient } from '@prisma/client';
-
-const prisma = new PrismaClient();
-export default prisma;
