--- libs/factory/src/lib/singelton.ts
@@ -1,15 +1,0 @@
-import { PrismaClient } from '@prisma/client';
-import { mockDeep, mockReset, DeepMockProxy } from 'jest-mock-extended';
-
-import prisma from './client';
-
-jest.mock('./client', () => ({
-  __esModule: true,
-  default: mockDeep<PrismaClient>(),
-}));
-
-beforeEach(() => {
-  mockReset(prismaMock);
-});
-
-export const prismaMock = prisma as unknown as DeepMockProxy<PrismaClient>;
